* arts-2-3-425
* CMakeLists.txt, CHANGES-2.3, README:

Don't enable NetCDF support by default. It's very limited and hardly
used. Can be enabled if needed with a cmake option.

* src/CMakeLists.txt:

Only add NetCDF source files to artscore lib if enabled.

* src/methods.cc: Don't mask % with \.

* src/m_tmatrix.cc (scat_data_singleTmatrix):

Add two explicit casts to Numeric to avoid compiler warning.

git-svn-id: https://www.sat.ltu.se/svn/rt/arts/trunk@9674 aaf1aab0-4228-0410-ad68-8dceda47f409

+++ src/m_tmatrix.cc
@@ -195,10 +195,10 @@ void scat_data_singleTmatrix(
           throw runtime_error( os.str() );
         }
       Index nza = data_za_grid.nelem();
-      Numeric dza = 180./(nza-1);
+      Numeric dza = 180./((Numeric)nza-1.);
       for( Index iza=1; iza<nza; iza++ )
         {
-          if( !(is_same_within_epsilon(data_za_grid[iza],iza*dza,2*DBL_EPSILON)) )
+          if( !(is_same_within_epsilon(data_za_grid[iza],(Numeric)iza*dza,2*DBL_EPSILON)) )
             {
               ostringstream os;
               os << "Input zenith angle grid *data_za_grid* is required to be\n"
